web/requirements.txt
Django==5.1.6
gunicorn
requests
django-dotenv
psycopg2-binary
django-storages
boto3
